Issue #4575: Modified BooleanExpressionComplexityCheckTest.java and moved its input files to the booleanexpressioncomplexity subdirectory
